Content of folder
The main file shows a grid of 36 questions where the user has to press on a number to show it, then has to select the possible answer showed with radio buttons, at the top a timer is showed. Also, this code has the functionality to generate an encrypted pdf file using gendPDF.php

This code includes: HTML, CSS, JavaScript, PHP and PHP libraries.